transcribe/requirements.txt
openai
python-dotenv
ffmpeg-python
openai-whisper